+++ app/addons/replication/components/replicate-activity.js
