+++ docker/dc.selenium.yml
@@ -2,13 +2,13 @@ version: '2'
 services:
   selenium:
     container_name: selenium
-    image: selenium/standalone-firefox:3.1.0
+    image: selenium/standalone-firefox:2.48.2
     ports:
       - "4444:4444"
 
   couchdb:
     container_name: couchdb
-    image: klaemo/couchdb:2.0-dev
+    image: klaemo/couchdb:2.0-dev@sha256:e9b71abaff6aeaa34ee28604c3aeb78f3a7c789ad74a7b88148e2ef78f1e3b21
     command: '--with-haproxy -a tester:testerpass'
     ports:
       - "5984:5984"
